+++ database/migrations/20240806195930_add_type_to_presences_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddTypeToPresencesTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('presences', function (Blueprint $table) {
+            $table->string('type', 12)->nullable()->index();
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('presences', function (Blueprint $table) {
+            $table->dropColumn('type');
+        });
+    }
+}