+++ src/Moxl/Xec/Action/Roster/RemoveItem.php
@@ -17,14 +17,11 @@ class RemoveItem extends Action
 
     public function handle(?\SimpleXMLElement $stanza = null, ?\SimpleXMLElement $parent = null)
     {
-        $contact = \App\User::me()
+        \App\User::me()
               ->session
               ->contacts()
               ->where('jid', $this->_to)
-              ->first();
-        if ($contact) {
-            $contact->delete();
-        }
+              ->delete();
 
         $this->pack($this->_to);
         $this->deliver();